libgimpbase: fix setting a GimpParamChoice to invalid value.
It was possible to set a GimpChoice property to a value out of the
allowed list. This got very likely broken when I re-parented
GimpParamSpecChoice to GParamSpecString which had a value_is_valid()
implementation, which was shadowing our custom value_validate().

This is fixed by now also implementing value_is_valid() for this param
spec type.